backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
numpy==2.1.2
scikit-learn==1.5.2
PyPDF2==3.0.1
PyJWT==2.9.0
